move types.py out of utils
